+++ _script/package.sh
@@ -35,9 +35,26 @@ mkproject() {
     echo
 }
 
+# Find texture names which file data ID is lesser or equal to a specific threshold
+# Such textures are supposed to be already embedded in the game files
+# $1 = threshold up until textures are supposed to be embedded, and can be pruned
+# $2+ = IDs to avoid even if they are below the threshold
+texbelow() {
+    local threshold=$1
+    shift
+    awk '/^local mapping/{flag=1;next;next} /^}/{flag=0} flag' SpellActivationOverlay/textures/texname.lua |
+        tr " \t" "_" | tr -d "'" |
+        awk -F'"' "1${*/#/ \&\& \$2!=}" |
+        awk -F'"' "{ if (\$2 <= $threshold) print \$4 }" |
+        while read name
+        do
+            [ -e "SpellActivationOverlay/textures/${name,,}.blp" ] && printf '%s\n' "${name,,}"
+        done
+}
+
 # Remove unused textures to reduce archive size.
 # The list passed as parameter is based on the contents of the array
-# SpellActivationOverlayDB.debug.unmarked after calling the global
+# SpellActivationOverlayDB.dev.unmarked after calling the global
 # function SAO_DB_ComputeUnmarkedTextures() on each and every class
 # on characters logged in with the game client of the target flavor.
 # Because these textures are not 'marked', we don't need them.
@@ -123,7 +140,8 @@ raging_blow
 arcane_missiles_1
 arcane_missiles_2
 arcane_missiles_3
-fulmination)
+fulmination
+sudden_doom)
 prunetex "${TEXTURES_NOT_FOR_WRATH[@]}"
 
 zipproject wrath "$VERSION_TOC_VERSION"
@@ -146,9 +164,9 @@ shooting_stars
 daybreak
 backlash
 predatory_swiftness
-sword_and_board
 killing_machine
-rime)
+rime
+sudden_doom)
 prunetex "${TEXTURES_NOT_FOR_VANILLA[@]}"
 
 zipproject vanilla "$VERSION_TOC_VERSION"
@@ -166,7 +184,9 @@ raging_blow
 arcane_missiles_1
 arcane_missiles_2
 arcane_missiles_3
-fulmination)
+fulmination
+$(texbelow 511469 450914 450915)
+)
 prunetex "${TEXTURES_NOT_FOR_CATA[@]}"
 
 SOUNDS_NOT_FOR_CATA=(UI_PowerAura_Generic)