+++ classes/hunter.lua
@@ -49,6 +49,9 @@ local function registerClass(self)
         -- Lock and Load (Season of Discovery)
         -- Unlike Wrath, we do not suggest to glow buttons, because there are too many (all 'shots')
         self:RegisterAura("lock_and_load", 0, 415414, "lock_and_load", "Top", 1, 255, 255, 255, true);
+
+        -- Sniper Training, 5 stacks only (Season of Discovery)
+        -- self:RegisterAura("sniper_training", 5, 415401, nil, "", 0, 0, 0, 0, false, { (GetSpellInfo(aimedShot)) }, true);
     end
 end
 
@@ -71,6 +74,8 @@ local function loadOptions(self)
 
     local flankingStrike = 415320;
     local cobraStrikes = 425714;
+    -- local sniperTrainingBuff = 415401;
+    -- local sniperTrainingRune = 415399;
 
     if self.IsWrath() or self.IsCata() then
         self:AddOverlayOption(improvedSteadyShotTalent, improvedSteadyShotBuff);
@@ -91,6 +96,7 @@ local function loadOptions(self)
     end
     if self.IsSoD() then
         self:AddGlowingOption(nil, flankingStrike, flankingStrike);
+        -- self:AddGlowingOption(sniperTrainingRune, sniperTrainingBuff, aimedShot, self:NbStacks(5));
     end
     if self.IsWrath() or self.IsCata() then
         self:AddGlowingOption(improvedSteadyShotTalent, improvedSteadyShotBuff, aimedShot);