+++ classes/paladin.lua
@@ -53,6 +53,28 @@ local function registerClass(self)
 
         -- Healing Trance / Soul Preserver
         self:RegisterAuraSoulPreserver("soul_preserver_paladin", 60513); -- 60513 = Paladin buff
+    elseif self.IsCata() then
+        local infusionOfLightBuff1 = 53672;
+        local infusionOfLightBuff2 = 54149;
+        local artOfWarBuff = 59578;
+
+        local divineLight = 82326;
+        local holyRadiance = 82327;
+        local infusionOfLightButtons = { flashOfLight, holyLight, divineLight, holyRadiance };
+
+        -- Add option links during registerClass(), not because loadOptions() which would be loaded only when the options panel is opened
+        -- Add option links before RegisterAura() calls, so that options they are used by initial triggers, if any
+        self:AddOverlayLink(infusionOfLightBuff2, infusionOfLightBuff1);
+        self:AddGlowingLink(infusionOfLightBuff2, infusionOfLightBuff1);
+
+        -- Art of War
+        self:RegisterAura("art_of_war", 0, artOfWarBuff, "art_of_war", "Left + Right (Flipped)", 1, 255, 255, 255, true, { exorcism });
+
+        -- Infusion of Light, 1/2 talent points
+        self:RegisterAura("infusion_of_light_low", 0, infusionOfLightBuff1, "surge_of_light", "Top (CW)", 1, 255, 255, 255, true, infusionOfLightButtons);
+
+        -- Infusion of Light, 2/2 talent points
+        self:RegisterAura("infusion_of_light_high", 0, infusionOfLightBuff2, "surge_of_light", "Top (CW)", 1, 255, 255, 255, true, infusionOfLightButtons);
     end
 end
 
@@ -71,7 +93,6 @@ local function loadOptions(self)
 
     if self.IsWrath() then
         local flashOfLight = 19750;
-        local exorcism = 879;
         local holyLight = 635;
 
 --        local infusionOfLightBuff1 = 53672;
@@ -90,6 +111,27 @@ local function loadOptions(self)
         self:AddGlowingOption(infusionOfLightTalent, infusionOfLightBuff2, holyLight);
         self:AddGlowingOption(artOfWarTalent, artOfWarBuff2, exorcism);
         self:AddGlowingOption(artOfWarTalent, artOfWarBuff2, flashOfLight);
+    elseif self.IsCata() then
+        local flashOfLight = 19750;
+        local holyLight = 635;
+        local divineLight = 82326;
+        local holyRadiance = 82327;
+
+--        local infusionOfLightBuff1 = 53672;
+        local infusionOfLightBuff2 = 54149;
+        local infusionOfLightTalent = 53569;
+
+        local artOfWarBuff = 59578;
+        local artOfWarTalent = 53486;
+
+        self:AddOverlayOption(infusionOfLightTalent, infusionOfLightBuff2, 0, self:RecentlyUpdated()); -- Updated 2024-04-30
+        self:AddOverlayOption(artOfWarTalent, artOfWarBuff);
+
+        self:AddGlowingOption(infusionOfLightTalent, infusionOfLightBuff2, flashOfLight);
+        self:AddGlowingOption(infusionOfLightTalent, infusionOfLightBuff2, holyLight);
+        self:AddGlowingOption(infusionOfLightTalent, infusionOfLightBuff2, divineLight);
+        self:AddGlowingOption(infusionOfLightTalent, infusionOfLightBuff2, holyRadiance);
+        self:AddGlowingOption(artOfWarTalent, artOfWarBuff, exorcism);
     end
 end